+++ src/resdk/resources/base.py
@@ -192,7 +192,7 @@ class BaseResource:
         if payload := {
             field.server_field: field.to_json(getattr(self, field_name))
             for field_name, field in self._get_fields(to_payload).items()
-            if field.changed(self)
+            if field.changed(self) and field._sends_payload_if_changed(self)
         }:
             self._update_fields(api_call(payload), data_source=DataSource.SERVER)